==> notebooks/2_transform_data.ipynb <==
import pandas as pd

data = pd.read_parquet('../data/raw/raw_taxi_data.parquet')
data.head()

data = data[['tpep_pickup_datetime','PULocationID']]
data.rename(columns={'tpep_pickup_datetime':'PU_datetime','PULocationID':'PU_location'},inplace=True)
data.head()

data = data[data['PU_datetime']>= '2022-01-01']
data = data[data['PU_datetime'] < '2022-02-01']
data.head()

data['PU_hours'] = data['PU_datetime'].dt.floor('H')

data

agg_rides = data.groupby(['PU_hours','PU_location']).size().reset_index()
agg_rides.rename(columns={0:'rides'},inplace=True)
agg_rides

#now add the mission rows in data
from tqdm import tqdm

def fill_missing_time_slots(agg_rides : pd.DataFrame) -> pd.DataFrame:
    location_ids = agg_rides['PU_location'].unique()
    output = pd.DataFrame()
    range = pd.date_range(agg_rides['PU_hours'].min(),agg_rides['PU_hours'].max(),freq='H')
    
    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides['PU_location']==location_id,['PU_hours','rides']]

        agg_rides_i.set_index('PU_hours',inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(range,fill_value=0)

        agg_rides_i['PU_location'] = location_id

        output = pd.concat([output,agg_rides_i])
        
    output = output.reset_index().rename(columns={'index': 'PU_hours'})
    return output


complete_data = fill_missing_time_slots(agg_rides=agg_rides)

complete_data.describe()

#using plotly
from typing import Optional, List

import plotly.express as px

def plot_graph(
        location : Optional[List[int]]
    ):
    rides_to_plot = complete_data[complete_data['PU_location'].isin(location)] if location else data
    
    fig = px.line(
        rides_to_plot,
        x='PU_hours',
        y='rides',
        color='PU_location',
        template=None
    )
    fig.show()

plot_graph([43])

complete_data.to_parquet('../data/transformed/transformed_complete_data.parquet'
)
