==> notebooks/4_load_and_validate_data.ipynb <==
import sys 
sys.path.insert(0, 'E:/repos/my_projects/taxi_demand_project')

%reload_ext autoreload
%autoreload 2
from src.data import load_raw_data, transform_to_ts_data
raw = load_raw_data(year=2022)


from src.data import load_raw_data, transform_to_ts_data, transform_ts_data_to_features_and_targets


raw.head()

transformed = transform_to_ts_data(raw)

from src.data import load_raw_data, transform_to_ts_data, transform_ts_data_to_features_and_targets

features,targets = transform_ts_data_to_features_and_targets(transformed, 24*28*1, 24)

print('no of features',features.shape)
print('no of targets',targets.shape)


from src.paths import TRANSFORMED_DATA_DIR

tabular = features
tabular['target_next_hour'] = targets

tabular.to_parquet(TRANSFORMED_DATA_DIR/ 'transformed_tabular_data.parquet')
