==> notebooks/6_baseline_model.ipynb <==
import sys
sys.path.insert(0,'E:/repos/my_projects/taxi_demand_project')
sys.path

%reload_ext autoreload
%autoreload 2

import numpy as np
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
from src.data_split import train_test_split
from sklearn.metrics import mean_absolute_error


ts_data = pd.read_parquet(TRANSFORMED_DATA_DIR/'transformed_tabular_data.parquet')

ts_data.head()

X_train,X_test,y_train,y_test= train_test_split(ts_data=ts_data,
                                cut_off_year=2022,
                                cut_off_month=6, cut_off_day= 1,
                                target_column='target_next_hour')


X_train['PU_hour'].max()

X_train

class BaselineModelPreviousHour:
    def fit(self, X_train:pd.DataFrame, y_train:pd.Series):
        pass
    def predict(self, X_test:pd.DataFrame) -> np.array:
        return X_test['rides_1_hr_before']

base_model_1 = BaselineModelPreviousHour()
predictions_1 = base_model_1.predict(X_test=X_test)
base_model_mae_1 =mean_absolute_error(y_true=y_test,y_pred=predictions_1)
base_model_mae_1

class BaselineModelPreviousWeek:
    def fit(self, X_train:pd.DataFrame, y_train:pd.Series):
        pass
    def predict(self, X_test:pd.DataFrame) -> np.array:
        return X_test[f'rides_{7*24}_hr_before']

base_model_2 = BaselineModelPreviousWeek()
predictions_2 = base_model_2.predict(X_test=X_test)

base_model_mae_2 =mean_absolute_error(y_true=y_test,y_pred=predictions_2)
print(base_model_mae_2)

class BaselineModelPreviousWeeklyAverage:
    def fit(self, X_train:pd.DataFrame, y_train:pd.Series):
        pass
    def predict(self, X_test:pd.DataFrame) -> np.array:
        return 0.25*(X_test[f'rides_{7*24}_hr_before']+\
                     X_test[f'rides_{14*24}_hr_before']+\
                     X_test[f'rides_{21*24}_hr_before']+\
                     X_test[f'rides_{28*24}_hr_before'])

base_model_3 = BaselineModelPreviousWeeklyAverage()
predictions_3 = base_model_3.predict(X_test=X_test)

base_model_mae_3 =mean_absolute_error(y_true=y_test,y_pred=predictions_3)
print(base_model_mae_3)
