==> notebooks/9_lightgbm_regressor_with_feature_engineering.ipynb <==
import sys
sys.path.insert(0,'E:/repos/my_projects/taxi_demand_project')
from src.paths import TRANSFORMED_DATA_DIR
import pandas as pd
from src.data_split import train_test_split
import lightgbm as lgbm
from sklearn.metrics import mean_absolute_error as mae
import numpy as np

ts_data = pd.read_parquet(TRANSFORMED_DATA_DIR/'transformed_tabular_data.parquet')
ts_data

X_train,X_test,y_train,y_test= train_test_split(ts_data=ts_data,
                                cut_off_year=2022,
                                cut_off_month=6, cut_off_day= 1,
                                target_column='target_next_hour')


def average_4_week_rides (X:pd.DataFrame) -> np.array:
    X['average_rides_last_4_weeks'] = 0.25*(X[f'rides_{7*24}_hr_before']+\
                     X[f'rides_{14*24}_hr_before']+\
                     X[f'rides_{21*24}_hr_before']+\
                     X[f'rides_{28*24}_hr_before'])
    return X

from sklearn.preprocessing import FunctionTransformer

add_average_4_week_rides = FunctionTransformer(
    average_4_week_rides, validate=False
)

add_average_4_week_rides.fit_transform(X_train)

from sklearn.base import BaseEstimator , TransformerMixin

class TemporalFeatureEngineering(BaseEstimator,TransformerMixin):
    def fit(self,X,y=None):
        return self
    
    def transform(self,X,y=None):
        X_=X.copy()

        X_['hour'] = X_['PU_hour'].dt.hour
        X_['day_of_the_week'] = X_['PU_hour'].dt.dayofweek

        return X_.drop(columns = ['PU_hour'])

add_temporal_data = TemporalFeatureEngineering()

add_temporal_data.fit_transform(X_train)

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_average_4_week_rides,
    add_temporal_data,
    lgbm.LGBMRegressor()
)

pipeline.fit(X_train,y_train)

pipeline_predictions = pipeline.predict(X_test)

lgbm_mae = mae(y_pred=pipeline_predictions,y_true= y_test)

lgbm_mae
