==> notebooks/10_hyperparameters_tuning.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import sys
sys.path.insert(0,'E:/repos/my_projects/taxi_demand_project')
from src.paths import TRANSFORMED_DATA_DIR
import pandas as pd
from src.data_split import train_test_split

ts_data = pd.read_parquet(TRANSFORMED_DATA_DIR/'transformed_tabular_data.parquet')
ts_data

X_train,X_test,y_train,y_test= train_test_split(ts_data=ts_data,
                                cut_off_year=2022,
                                cut_off_month=6, cut_off_day= 1,
                                target_column='target_next_hour')


from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline,Pipeline
from sklearn.metrics import mean_absolute_error as mae
import optuna
import numpy as np
from src.model import training_pipeline

def objective(trial : optuna.trial.Trial)-> float:

    objective_hyperparameters= {
        "metric": 'mae',
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100), 
    }

    ts_splits = TimeSeriesSplit(n_splits=2)
    scores = []
    for train_split, val_split in ts_splits.split(X_train):
        
        X_train_, X_val_ = X_train.iloc[train_split,:], X_train.iloc[val_split,:]
        y_train_, y_val_ = y_train.iloc[train_split], y_train.iloc[val_split]

        pipeline = training_pipeline(**objective_hyperparameters)
        pipeline.fit(X_train_,y_train_)

        preds = pipeline.predict(X_val_)

        split_mae = mae(y_pred=preds,y_true=y_val_)

        scores.append(split_mae)

    return np.array(scores).mean()

study = optuna.create_study(direction='minimize')
study.optimize(objective,n_trials=4)

params = study.best_trial.params
best_pipeline = training_pipeline(**params)

best_pipeline.fit(X_train,y_train)

preds= best_pipeline.predict(X_test)

print(mae(y_true=y_test,y_pred=preds))

%reload_ext autoreload
%autoreload 2
from src.plot import plot_one_example



plot_one_example(X_test,
                       y_test,
                       18940,
                       prediction=pd.Series(preds))
