==> notebooks/11_backfill_feature_store.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys 
sys.path.insert(0, 'E:/repos/my_projects/taxi_demand_predictor')

from dotenv import load_dotenv
import os
from src.paths import PATRENT_DIR

HOPSWORKS_PROJECT = 'taxi_demand_salar'

load_dotenv(PATRENT_DIR/'.env')
HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

from datetime import datetime
from src.data import load_raw_data , transform_to_ts_data
import pandas as pd



from_date = 2022
end_date = datetime.now().year

print(f'downloading data from {from_date} to {end_date}')
rides_data =pd.DataFrame()

for year in range(from_date,end_date+1):
    rides_one_year = load_raw_data(year=year)

    rides_data = pd.concat([rides_data,rides_one_year])




rides_data.shape


rides_data.head()

rides_data = transform_to_ts_data(raw_data=rides_data)

rides_data.head()


import hopsworks

project = hopsworks.login(
    project= HOPSWORKS_PROJECT,
    api_key_value= HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'ts_data_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name = FEATURE_GROUP_NAME,
    version= FEATURE_GROUP_VERSION,
    description= "Time Series Data with Hourly frequency",
    primary_key= ['pu_location','pu_hour'],
    event_time= 'pu_hour'
)

feature_group.insert(rides_data, write_options={'wait_for_job':False})
