==> notebooks/12_feature_pipeline.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys 
sys.path.insert(0, 'E:/repos/NYC_tax_demand_predictor')

from datetime import timedelta,datetime
from src.data import load_raw_data,transform_to_ts_data
import pandas as pd

current_time = pd.to_datetime(datetime.utcnow()).floor('H')
print(f'current time {current_time}')

fetch_data_from = current_time- timedelta(days=28)
fetch_data_to = current_time

def fetch_data(
        data_from : datetime,
        data_to : datetime
)-> pd.DataFrame:

    fetch_data_from_ = data_from - timedelta(days=7*52)
    fetch_data_to_ = data_to - timedelta(days=7*52)

    rides = load_raw_data(fetch_data_from_.year,fetch_data_from_.month)
    rides = rides[rides['pu_datetime'] >= fetch_data_from_]
    rides2 = load_raw_data(fetch_data_to_.year,fetch_data_to_.month)
    rides2 = rides2[rides2['pu_datetime']<fetch_data_to_]

    rides = pd.concat([rides,rides2])
    rides['pu_datetime'] += timedelta(days=7*52)

    rides.sort_values(by=['pu_location','pu_datetime'],inplace= True)

    return rides

historical_data = fetch_data(data_from=fetch_data_from,data_to=fetch_data_to)

historical_data.shape

historical_data.head()

transformed_data = transform_to_ts_data(historical_data)

transformed_data.shape

transformed_data.head()

import hopsworks
import src.config as config

print(config.HOPSWORKS_PROJECT)
print(config.HOPSWORKS_API_KEY)

project = hopsworks.login(
    project= config.HOPSWORKS_PROJECT,
    api_key_value= config.HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

feature_group = feature_store.get_or_create_feature_group(
    name = config.FEATURE_GROUP_NAME,
    version= config.FEATURE_GROUP_VERSION,
    description= 'Time series data with hourly frequency',
    primary_key= ['pu_location','pu_hour'],
    event_time='pu_hour'
)

feature_group.insert(transformed_data, write_options={'wait_for_job':False})
